+++ packages/motherduck/src/index.ts
@@ -1,0 +1,11 @@
+/**
+ * {@include ../README.md}
+ * @packageDocumentation
+ */
+
+export {
+  createWasmMotherDuckDbConnector,
+  isWasmMotherDuckDbConnector,
+  type WasmMotherDuckDbConnector,
+  type WasmMotherDuckDbConnectorOptions,
+} from './WasmMotherDuckDbConnector';